requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
pypdf
pdfplumber
streamlit
ollama
tqdm
python-dotenv
scikit-learn
pandas
numpy
langchain
langchain-community
sentence-transformers
faiss-cpu
pdfplumber
streamlit
beautifulsoup4
requests
python-dotenv
tqdm
